Revert "feat: [#191] Add user actions"
This reverts commit 87342099fc4ed5f093123e53a4f71f7b3d1b7069.
