+++ backend/app/schemas/session_data.py
@@ -8,4 +8,3 @@ class SessionData(BaseModel):
     username: Optional[str]
     aai_state: Optional[str]
     rp_handler: Any
-    session_uuid: str
