+++ ui/apps/ui/src/app/components/results-with-pagination/results-with-pagination.component.ts
@@ -46,7 +46,6 @@ import { Router } from '@angular/router';
       <ess-result
         class="results"
         *ngFor="let result of results$ | async"
-        [id]="result.id"
         [title]="result.title"
         [description]="result.description"
         [type]="result.type"
